+++ app/src/main/java/com/example/androiddevchallenge/ui/navigation/AdoptyNavigation.kt
@@ -1,0 +1,64 @@
+/*
+ * MIT License
+ *
+ * Copyright (c) 2021 Vivek Singh
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+package com.example.androiddevchallenge.ui.navigation
+
+import androidx.compose.runtime.Composable
+import androidx.core.splashscreen.SplashScreen
+import androidx.lifecycle.viewmodel.compose.viewModel
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import androidx.navigation.compose.rememberNavController
+import com.example.androiddevchallenge.ui.screen.dogDetails.DogDetailScreen
+import com.example.androiddevchallenge.ui.screen.home.HomeScreen
+import com.example.androiddevchallenge.ui.screen.home.HomeViewModel
+
+@Composable
+fun AdoptyNavigation(splashScreenVisibleCondition: (SplashScreen.KeepOnScreenCondition) -> Unit) {
+    val navController = rememberNavController()
+    val viewModel: HomeViewModel = viewModel()
+
+    splashScreenVisibleCondition {
+        viewModel.state.value.run {
+            petState.loading || specialNeedsDogState.loading
+        }
+    }
+
+    NavHost(
+        navController = navController,
+        startDestination = Screen.Home.route,
+    ) {
+        composable(Screen.Home.route) {
+            HomeScreen(
+                viewModel = viewModel,
+                navController = navController,
+            )
+        }
+        composable(Screen.DogDetail.route) {
+            DogDetailScreen(
+                viewModel = viewModel,
+                navController = navController
+            )
+        }
+    }
+}
